Created functionality for displaying a div
Added functionality so that certain points aren't visible if you haven't cleared that phase yet and created a div that shows up only when you click on one of the locations. We haven't yet figured out how to dipslay the div if you click on the same point several times.

Co-Authored-By: Sami <83646827+samellow-afk@users.noreply.github.com>