Dialogue-related changes
Changed the names of the background-images to remove the unnecessary _klar-parts. Added a background-image for Kockums. Made it so that dialog-init works and rerouted the go-button to call on that instead. Added functionality so that the respective background is displayed during dialogue and that if you're not in a dialogue the overlay is instead displayed.

Yet another change is that if the name of the dialogue-obj is either BSK or Jag a completley transparent image is shown and if it is specifically BSK the name-sign is also hidden.