BSK with images now works
All character images now works. I simply replaced all empty image-keys with strings that directs to empty.png and then made it so that if the name is BSK the only difference is that the name-sign is hidden. In all other cases the name-sign is visible and in all cases the character-image that is displayed is the one in the image-key.