Fixed the map-points appearing
Changed it so that the map-points doesn't all appear when you, for example, start a dialogue. Instead of toggling the none-class when the id-number is greater than the current phase it now instead adds the class when the id-number is greater and removes it when it is less than the current phase.