Piano css and map-points
Changed the css and text of the piano game's game over screens so that they are more in line with the graphic profile. I also fixed the problem with the map points being visible for a split second when loading the page by making all of them have display: none at initialization.